requirements.txt
humanfriendly==9.1
gunicorn==20.1.0
Flask-Dance==3.3.1
Authlib==0.15.3
python-dotenv==0.15.0
pymongo==3.11.3
sendgrid==6.6.0
eventlet==0.30.1
python-dotenv==0.15.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
callofduty.py==1.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.5
dnspython==1.16.0
eventlet==0.30.1
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SocketIO==4.3.1
gevent==21.1.2
greenlet==1.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
linecache2==1.0.0
MarkupSafe==1.1.1
mongoengine==0.23.0
mongomock==3.22.1
multidict==5.1.0
packaging==20.9
passlib==1.7.4
Pillow==8.1.2
pluggy==0.13.1
py==1.10.0
pymongo==3.11.3
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==3.13.2
python-http-client==3.3.2
python-socketio==4.6.0
requests==2.25.1
rfc3986==1.4.0
sendgrid==6.6.0
sentinels==1.0.0
six==1.15.0
sniffio==1.2.0
starkbank-ecdsa==1.1.0
toml==0.10.2
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.3
valorant-api==0.0.4
Werkzeug==1.0.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
